Get broken tests to work on LLVM

+++ runtime/column_index.h
@@ -14,16 +14,16 @@
 namespace kush::runtime::ColumnIndex {
 
 // Create the column index
-absl::flat_hash_map<int8_t, std::vector<int32_t>>* CreateInt8Index();
-absl::flat_hash_map<int16_t, std::vector<int32_t>>* CreateInt16Index();
+absl::flat_hash_map<int32_t, std::vector<int32_t>>* CreateInt8Index();
+absl::flat_hash_map<int32_t, std::vector<int32_t>>* CreateInt16Index();
 absl::flat_hash_map<int32_t, std::vector<int32_t>>* CreateInt32Index();
 absl::flat_hash_map<int64_t, std::vector<int32_t>>* CreateInt64Index();
 absl::flat_hash_map<double, std::vector<int32_t>>* CreateFloat64Index();
 absl::flat_hash_map<std::string, std::vector<int32_t>>* CreateTextIndex();
 
 // Free the column index
-void FreeInt8Index(absl::flat_hash_map<int8_t, std::vector<int32_t>>* index);
-void FreeInt16Index(absl::flat_hash_map<int16_t, std::vector<int32_t>>* index);
+void FreeInt8Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index);
+void FreeInt16Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index);
 void FreeInt32Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index);
 void FreeInt64Index(absl::flat_hash_map<int64_t, std::vector<int32_t>>* index);
 void FreeFloat64Index(absl::flat_hash_map<double, std::vector<int32_t>>* index);
@@ -31,9 +31,9 @@ void FreeTextIndex(
     absl::flat_hash_map<std::string, std::vector<int32_t>>* index);
 
 // Insert tuple idx
-void InsertInt8Index(absl::flat_hash_map<int8_t, std::vector<int32_t>>* index,
+void InsertInt8Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index,
                      int8_t value, int32_t tuple_idx);
-void InsertInt16Index(absl::flat_hash_map<int16_t, std::vector<int32_t>>* index,
+void InsertInt16Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index,
                       int16_t value, int32_t tuple_idx);
 void InsertInt32Index(absl::flat_hash_map<int32_t, std::vector<int32_t>>* index,
                       int32_t value, int32_t tuple_idx);
@@ -48,9 +48,9 @@ void InsertTextIndex(
 
 // Get the bucket of tuple idx from index
 std::vector<int32_t>* GetBucketInt8Index(
-    absl::flat_hash_map<int8_t, std::vector<int32_t>>* index, int8_t value);
+    absl::flat_hash_map<int32_t, std::vector<int32_t>>* index, int8_t value);
 std::vector<int32_t>* GetBucketInt16Index(
-    absl::flat_hash_map<int16_t, std::vector<int32_t>>* index, int16_t value);
+    absl::flat_hash_map<int32_t, std::vector<int32_t>>* index, int16_t value);
 std::vector<int32_t>* GetBucketInt32Index(
     absl::flat_hash_map<int32_t, std::vector<int32_t>>* index, int32_t value);
 std::vector<int32_t>* GetBucketInt64Index(
